Reworked IO of IM and NM
The client and servers are now reworked to use async technology. The instance manager works as an async server. The NM has an async client. Hearbeats send by NM are, currently, replied with the same heartbeat to acknowledge. Later this could also be a command. The client could buffer messages, sends them all and each time waits for a response.

TODO: functionality for received commands on the client side
TODO: Process heartbeats on the server side