Created async function for filling taskpool
The create_full_taskpool now reads in all tasks into a list, sorts them on time and adds them whenever the time is equal to the current_time. There is a sleep(1) so it can wait a second in between before adding a new task. The self_tasks is now filled with all tasks, so these can be divided over workers.

I could not test it on Input.csv as I do not have this file. Note that you can have multiple jobs at the same time. Time is assumed to start at 0 and then goes to 1, 2, 3. You can have skips here. Task 1 on time 0 and task 2 on time 20, means task 1 is added immediately, Task 2 is added after 20 seconds.
Maybe you want to give it some time during the early testing phase.

See comment `run_task_pool`  for more information on how to actuall divide the work to workers